Rename bootstrap.ts naming fossil to message-sanitization.ts
Closes #264
