+++ src/agents/agent-helpers/google.ts
@@ -1,4 +1,4 @@
-import { sanitizeGoogleTurnOrdering } from "./bootstrap.js";
+import { sanitizeGoogleTurnOrdering } from "./message-sanitization.js";
 
 export function isGoogleModelApi(api?: string | null): boolean {
   return api === "google-gemini-cli" || api === "google-generative-ai";
